+++ dsr_rqt_plugin/src/dsr_rqt_plugin/dsr_rqt_plugin.cpp
@@ -15,36 +15,43 @@
 
 namespace dsr_rqt_plugin{
 
-DSRView::DSRView(): rqt_gui_cpp::Plugin(), widget_(0){
-	//setObjectName("DSRView");
+DSRView::DSRView(): rqt_gui_cpp::Plugin(), widget_graph_(0), widget_tree_(0){
 }
 
 DSRView::~DSRView(){
-
 }
 
 /* Initialize the publishers and subscribers */
 void DSRView::initPlugin(qt_gui_cpp::PluginContext& context){
-	// TODO: Fix this
+	// Start DSR graph
 	G_ = std::make_shared<DSR::DSRGraph>(0, "DSRView", 255, "");
-	widget_ = new QWidget();
-	QMainWindow *window;
-	graph_viewer_ = std::make_unique<DSR::DSRViewer>(window, G_, DSR::DSRViewer::view::tree, DSR::DSRViewer::view::graph);
-	window->setWindowFlags(Qt::Widget | Qt::FramelessWindowHint | Qt::WindowStaysOnTopHint | Qt::Tool);
-	context.addWidget(window);
-
-	/*auto centralwidget = new QMainWindow(this);
-	centralwidget->setWindowFlags(Qt::Widget);
-	setCentralWidget(centralwidget);*/
+	// Set the DSR viewer
+	QMainWindow *window = new QMainWindow();
+	int current_opts = DSR::DSRViewer::view::tree | DSR::DSRViewer::view::graph | 
+		DSR::DSRViewer::view::scene | DSR::DSRViewer::view::osg;
+	graph_viewer_ = std::make_unique<DSR::DSRViewer>(
+		window, G_, current_opts, DSR::DSRViewer::view::graph);
+	// FIXME: This is a hack to avoid the window to be shown
+	window->setVisible(false);
+	// Add the widgets to the user interface
+	widget_graph_ = graph_viewer_->get_widget(DSR::DSRViewer::view::graph);
+	widget_graph_->setWindowTitle("DSRGraph");
+	context.addWidget(widget_graph_);
+	widget_tree_ = graph_viewer_->get_widget(DSR::DSRViewer::view::tree);
+	widget_tree_->setWindowTitle("DSRTree");
+	context.addWidget(widget_tree_);
 }
 
 void DSRView::shutdownPlugin(){
+	G_.reset();
 }
 
-void DSRView::saveSettings(qt_gui_cpp::Settings& plugin_settings, qt_gui_cpp::Settings& instance_settings) const{
+void DSRView::saveSettings(qt_gui_cpp::Settings& plugin_settings, 
+	qt_gui_cpp::Settings& instance_settings) const{
 }
 
-void DSRView::restoreSettings(const qt_gui_cpp::Settings& plugin_settings, const qt_gui_cpp::Settings& instance_settings){
+void DSRView::restoreSettings(const qt_gui_cpp::Settings& plugin_settings, 
+	const qt_gui_cpp::Settings& instance_settings){
 }
 }